requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
requests>=2.25.0
matplotlib>=3.5.0
seaborn>=0.11.0
numpy>=1.21.0